Adding configuration ParameterMapper in FieldsBuilder
